+++ CHANGELOG.md
@@ -21,3 +21,10 @@
 - Added method toList for the Color object
 - Update docs for constants in presets library
 - Added method toMap for the Color object
+
+## 1.0.0
+
+- Change self-enum color components with Dart builtin enums
+- Remove unused enum items (RGB, RGBA, HEX)
+- Move color components from `package:fff/parser.dart` to `package:fff/color.dart`
+- Restore of the tests for work with new Dart test library